refactor(core): Standardize all IDs to use Long type instead of int

+++ src/main/java/com/example/personal_expense_manager/dto/request/TransactionRequest.java
@@ -22,5 +22,5 @@ public class TransactionRequest {
     String description;
     LocalDate date;
     TransactionType type;
-    int categoryId;     
+    Long categoryId;     
 }
